requirements.txt
streamlit==1.45.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
xgboost==2.1.4
plotly==6.0.1
joblib==1.4.2
PyYAML==6.0.2
imbalanced-learn==0.13.0
